added go.obo file
"go.obo and go.owl
These files contain the core GO ontology in two formats, OBO and
OWL-RDF/XML. This view includes relationships not in the filtered
version of GO including has_part and occurs_in. The set of relationships
used is likely to extend over time. Many of these relationships may not
be safe for propagating annotations across, so this version should not
be used with legacy GO tools. This version currently excludes
relationships to external ontologies."